Agora.Effect

Description

Helpers for constructing effects.

Documentation

makeEffect :: forall (datum :: PType). (PIsData datum, PTryFrom PData datum) => CurrencySymbol -> (forall (s :: S). Term s PCurrencySymbol -> Term s datum -> Term s PTxOutRef -> Term s (PAsData PTxInfo) -> Term s POpaque) -> ClosedTerm PValidator

Helper "template" for creating effect validator.

In some situations, it may be the case that we need more control over how an effect is implemented. In such situations, it's okay to not use this helper.