Agora.Effect.TreasuryWithdrawal

Description

An Effect that withdraws treasury deposit

Documentation

data TreasuryWithdrawalDatum

Datum that encodes behavior of Treasury Withdrawal effect.

Note: This Datum acts like a "predefined redeemer". Which is to say that it encodes the properties a redeemer would, but is locked in-place until spend.

Constructors

TreasuryWithdrawalDatum

Fields

receivers :: [(Credential, Value)]

AssocMap for Value sent to each receiver from the treasury.

treasuries :: [Credential]

What Credentials is spending from legal.

newtype PTreasuryWithdrawalDatum (s :: S)

Haskell-level version of TreasuryWithdrawalDatum.

Constructors

PTreasuryWithdrawalDatum (Term s (PDataRecord '["receivers" := PBuiltinList (PAsData (PTuple PCredential (PValue 'Sorted 'Positive))), "treasuries" := PBuiltinList (PAsData PCredential)]))

treasuryWithdrawalValidator :: forall {s :: S}. CurrencySymbol -> Term s PValidator

Withdraws given list of values to specific target addresses. It can be evoked by burning GAT. The transaction should have correct outputs to the users and any left overs should be paid back to the treasury.

The validator does not accept any Redeemer as all "parameters" are provided via encoded Datum.

Note: It should check... 1. Transaction outputs should contain all of what Datum specified 2. Left over assets should be redirected back to Treasury It can be more flexiable over... - The number of outputs themselves