Agora.MultiSig

Description

A basic N of M multisignature validation function.

Documentation

validatedByMultisig :: MultiSig -> Term s (PTxInfo :--> PBool)

Check if a Haskell-level MultiSig signs this transaction.

pvalidatedByMultisig :: Term s (PMultiSig :--> (PTxInfo :--> PBool))

Check if a Plutarch-level MultiSig signs this transaction.

newtype PMultiSig (s :: S)

Plutarch-level MultiSig

Constructors

PMultiSig

Fields

getMultiSig :: Term s (PDataRecord '["keys" := PBuiltinList (PAsData PPubKeyHash), "minSigs" := PInteger])

data MultiSig

A MultiSig represents a proof that a particular set of signatures are present on a transaction.

Constructors

MultiSig

Fields

keys :: [PubKeyHash]

List of PubKeyHashes that must be present in the list of signatories.

minSigs :: Integer