Agora.Proposal

Description

Proposal scripts encoding effects that operate on the system.

Haskell-land

data Proposal

Parameters that identify the Proposal validator script.

Constructors

Proposal

data ProposalDatum

Haskell-level datum for Proposal scripts.

Constructors

ProposalDatum

Fields

effects :: Map ResultTag [(ValidatorHash, DatumHash)]

Effect lookup table. First by result, then by effect hash.

status :: ProposalStatus

The status the proposal is in.

cosigners :: [PubKeyHash]

Who created the proposal initially, and who cosigned it later.

thresholds :: ProposalThresholds

Thresholds copied over on initialization.

votes :: ProposalVotes

Vote tally on the proposal

data ProposalStatus

The "status" of the proposal. This is only useful for state transitions, as opposed to time-based "phases".

If the proposal is VotingReady, for instance, that doesn't necessarily mean that voting is possible, as this also requires the timing to be right.

Constructors

Draft
A draft proposal represents a proposal that has yet to be realized.

In effect, this means one which didn't have enough LQ to be a full proposal, and needs cosigners to enable that to happen. This is similar to a "temperature check", but only useful if multiple people want to pool governance tokens together. If the proposal doesn't get to VotingReady on time, the proposal will never be able to get voted on.

VotingReady
The proposal has/had enough GT cosigned in order to be a fully fledged proposal.

This means that once the timing requirements align, proposal will be able to be voted on.

Finished
The proposal has finished.

This can mean it's been voted on and completed, but it can also mean the proposal failed due to time constraints or didn't get to VotingReady first.

TODO: The owner of the proposal may choose to reclaim their proposal.

data ProposalThresholds

The threshold values for various state transitions to happen. This data is stored centrally (in the Governor) and copied over to Proposals when they are created.

Constructors

ProposalThresholds

Fields

execute :: Tagged GTTag Integer

How much GT minimum must a particular ResultTag accumulate for it to pass.

draft :: Tagged GTTag Integer

How much GT required to "create" a proposal.

vote :: Tagged GTTag Integer

How much GT required to allow voting to happen. (i.e. to move into VotingReady)

newtype ProposalVotes

Map which encodes the total tally for each result. It's important that the "shape" is consistent with the shape of effects.

e.g. if the effects field looks like the following:

[(ResultTag 0, []), (ResultTag 1, [(vh, dh)])]

Then ProposalVotes needs be of the shape:

[(ResultTag 0, n), (ResultTag 1, m)]

Constructors

ProposalVotes

Fields

getProposalVotes :: Map ResultTag Integer

newtype ProposalId

Identifies a Proposal, issued upon creation of a proposal. In practice, this number starts at zero, and increments by one for each proposal. The 100th proposal will be ProposalId 99. This counter lives in the Governor, see nextProposalId.

Constructors

ProposalId

Fields

proposalTag :: Integer

newtype ResultTag

Encodes a result. Typically, for a Yes/No proposal, we encode it like this:

No  ~ ResultTag 0
Yes ~ ResultTag 1

Constructors

ResultTag

Fields

getResultTag :: Integer

Plutarch-land

newtype PProposalDatum (s :: S)

Plutarch-level version of ProposalDatum.

Constructors

PProposalDatum

Fields

getProposalDatum :: Term s (PDataRecord '["effects" := PMap PResultTag (PMap PValidatorHash PDatumHash), "status" := PProposalStatus, "cosigners" := PBuiltinList PPubKeyHash, "thresholds" := PProposalThresholds, "votes" := PProposalVotes])

data PProposalStatus (s :: S)

Plutarch-level version of ProposalStatus.

Constructors

PDraft (Term s (PDataRecord '[]))

PVotingReady (Term s (PDataRecord '[]))

PFinished (Term s (PDataRecord '[]))

newtype PProposalThresholds (s :: S)

Plutarch-level version of ProposalThresholds.

Constructors

PProposalThresholds

Fields

getProposalThresholds :: Term s (PDataRecord '["execute" := PDiscrete GTTag, "draft" := PDiscrete GTTag, "vote" := PDiscrete GTTag])

newtype PProposalVotes (s :: S)

Plutarch-level version of ProposalVotes.

Constructors

PProposalVotes (Term s (PMap PResultTag PInteger))

newtype PProposalId (s :: S)

Plutarch-level version of PProposalId.

Constructors

PProposalId (Term s PInteger)

newtype PResultTag (s :: S)

Plutarch-level version of ResultTag.

Constructors

PResultTag (Term s PInteger)

Scripts

proposalValidator :: Proposal -> ClosedTerm PValidator

Validator for Proposals.

proposalPolicy :: Proposal -> ClosedTerm PMintingPolicy

Policy for Proposals.