Agora.Proposal.Time

Description

Time functions for proposal phases.

Haskell-land

data ProposalTimingConfig

Configuration of proposal timings.

See: https://github.com/Liqwid-Labs/agora/blob/master/docs/tech-design/proposals.md#when-may-interactions-occur

Constructors

ProposalTimingConfig

Fields

draftTime :: POSIXTime

D: the length of the draft period.

votingTime :: POSIXTime

V: the length of the voting period.

lockingTime :: POSIXTime

L: the length of the locking period.

executingTime :: POSIXTime

E: the length of the execution period.

newtype ProposalStartingTime

Represents the starting time of the proposal.

Constructors

ProposalStartingTime

Fields

getProposalStartingTime :: POSIXTime

newtype MaxTimeRangeWidth

Represents the maximum width of a POSIXTimeRange.

Constructors

MaxTimeRangeWidth

Fields

getMaxWidth :: POSIXTime

Plutarch-land

data PProposalTime (s :: S)

Establishing timing in Proposal interactions.

In Plutus, it's impossible to determine time exactly. It's also impossible to get a single point in time, yet often we need to check various constraints on time.

For the purposes of proposals, there's a single most important feature: The ability to determine if we can perform an action. In order to correctly determine if we are able to perform certain actions, we need to know what time it roughly is, compared to when the proposal was created.

PProposalTime represents "the time according to the proposal". Its representation is opaque, and doesn't matter.

Various functions work simply on PProposalTime and ProposalTimingConfig. In particular, currentProposalTime is useful for extracting the time from the txInfoValidPeriod field of TxInfo.

We avoid PPOSIXTimeRange where we can in order to save on operations.

Note: PProposalTime doesn't need a Haskell-level equivalent because it is only used in scripts, and does not go in datums. It is also scott-encoded which is more efficient in usage.

Constructors

PProposalTime

Fields

lowerBound :: Term s PPOSIXTime

upperBound :: Term s PPOSIXTime

newtype PProposalTimingConfig (s :: S)

Plutarch-level version of ProposalTimingConfig.

Constructors

PProposalTimingConfig

Fields

getProposalTimingConfig :: Term s (PDataRecord '["draftTime" := PPOSIXTime, "votingTime" := PPOSIXTime, "lockingTime" := PPOSIXTime, "executingTime" := PPOSIXTime])

newtype PProposalStartingTime (s :: S)

Plutarch-level version of ProposalStartingTime.

Constructors

PProposalStartingTime (Term s PPOSIXTime)

newtype PMaxTimeRangeWidth (s :: S)

Plutarch-level version of MaxTimeRangeWidth.

Constructors

PMaxTimeRangeWidth (Term s PPOSIXTime)

Compute periods given config and starting time.

createProposalStartingTime :: forall (s :: S). Term s (PMaxTimeRangeWidth :--> (PPOSIXTimeRange :--> PProposalStartingTime))

Get the starting time of a proposal, from the txInfoValidPeriod field. For every proposal, this is only meant to run once upon creation. Given time range should be tight enough, meaning that the width of the time range should be less than the maximum value.

currentProposalTime :: forall (s :: S). Term s (PPOSIXTimeRange :--> PProposalTime)

Get the current proposal time, from the txInfoValidPeriod field.

If it's impossible to get a fully-bounded time, (e.g. either end of the PPOSIXTimeRange is an infinity) then we error out.

isDraftPeriod :: forall (s :: S). Term s (PProposalTimingConfig :--> (PProposalStartingTime :--> (PProposalTime :--> PBool)))

True if the PProposalTime is in the draft period.

isVotingPeriod :: forall (s :: S). Term s (PProposalTimingConfig :--> (PProposalStartingTime :--> (PProposalTime :--> PBool)))

True if the PProposalTime is in the voting period.

isLockingPeriod :: forall (s :: S). Term s (PProposalTimingConfig :--> (PProposalStartingTime :--> (PProposalTime :--> PBool)))

True if the PProposalTime is in the locking period.

isExecutionPeriod :: forall (s :: S). Term s (PProposalTimingConfig :--> (PProposalStartingTime :--> (PProposalTime :--> PBool)))

True if the PProposalTime is in the execution period.

Orphan instances

AdditiveSemigroup (Term s PPOSIXTime)

Instance details

Methods

(+) :: Term s PPOSIXTime -> Term s PPOSIXTime -> Term s PPOSIXTime