Agora.Record

Description

PDataRecord helper functions.

Documentation

mkRecord :: forall (r :: [PLabeledType]) (s :: S). RecordMorphism s '[] r -> Term s (PDataRecord r)

Turn a constant RecordMorphism into a fully built PDataRecord.

mkRecordConstr

Arguments

:: forall (r :: [PLabeledType]) (s :: S) (pt :: PType). PlutusType pt

=> (forall s'. Term s' (PDataRecord r) -> pt s')
The constructor. This is just the Haskell-level constructor for the type. For PMaybeData, this would be PDJust, or PNothing.

-> RecordMorphism s '[] r
The morphism that builds the record.

-> Term s pt

mkRecord but for known data-types.

This allows you to dynamically construct a record type constructor.

Example:

mkRecordConstr
  PStakeDatum
  ( #stakedAmount .= pconstantData (Tagged @GTTag 42)
      .& #owner .= pconstantData "aabbcc"
      .& #lockedBy .= pdata pnil
  )

Is the same as

pconstant (StakeDatum (Tagged 42) "aabbcc" [])

(.=) infix 7

Arguments

:: forall (sym :: Symbol) (a :: PType) (as :: [PLabeledType]) (s :: S). FieldName sym
The field name. You can use -XOverloadedLabels to enable the syntax: #hello ~ FieldName "hello"

-> Term s (PAsData a)
The value at that field. This must be PAsData, because the underlying type is Constr Integer [Data].

-> RecordMorphism s as ((sym := a) ': as)

Cons a labeled type as a RecordMorphism.

(.&) :: forall (s :: S) (a :: [PLabeledType]) (b :: [PLabeledType]) (c :: [PLabeledType]). RecordMorphism s b c -> RecordMorphism s a b -> RecordMorphism s a c infixr 6

Compose two RecordMorphisms.

data RecordMorphism (s :: S) (as :: [PLabeledType]) (bs :: [PLabeledType])

A morphism from one PDataRecord to another, representing some sort of consing of data.

data FieldName (sym :: Symbol)

Like Proxy but local to this module.