Agora.Stake

Description

Vote-lockable stake UTXOs holding GT.

Haskell-land

data StakeDatum

Haskell-level datum for Stake scripts.

Constructors

StakeDatum

Fields

stakedAmount :: Tagged GTTag Integer

Tracks the amount of governance token staked in the datum. This also acts as the voting weight for Proposals.

owner :: PubKeyHash

The hash of the public key this stake belongs to.

TODO Support for MultiSig/Scripts is tracked here: https://github.com/Liqwid-Labs/agora/issues/45

lockedBy :: [ProposalLock]

The current proposals locking this stake. This field must be empty for the stake to be usable for deposits and withdrawals.

data StakeRedeemer

Haskell-level redeemer for Stake scripts.

Constructors

DepositWithdraw (Tagged GTTag Integer)
Deposit or withdraw a discrete amount of the staked governance token. Stake must be unlocked.

Destroy
Destroy a stake, retrieving its LQ, the minimum ADA and any other assets. Stake must be unlocked.

PermitVote ProposalLock
Permit a Vote to be added onto a Proposal. This also adds a lock to the lockedBy field. See ProposalLock. This needs to be done in sync with casting a vote, otherwise it's possible for a lock to be permanently placed on the stake, and then the funds are lost.

RetractVotes [ProposalLock]
Retract a vote, removing it from the lockedBy field. See ProposalLock. This action checks for permission of the Proposal. Finished proposals are always allowed to have votes retracted and won't affect the Proposal datum, allowing Stakes to be unlocked.

WitnessStake
The owner can consume stake if nothing is changed about it. If the proposal token moves, this is equivalent to the owner consuming it.

data Stake

Parameters for creating Stake scripts.

Constructors

Stake

Fields

gtClassRef :: Tagged GTTag AssetClass

Used when inlining the AssetClass of a PDiscrete in the script code.

proposalSTClass :: AssetClass

data ProposalLock

A lock placed on a Stake datum in order to prevent depositing and withdrawing when votes are in place.

NOTE: Due to retracting votes always being possible, this lock will only lock with contention on the proposal.

FIXME: Contention on Proposals could create contention on voting which in turn creates contention on stakers.

Vaguely this is the dependency graph for this locking interaction. Both the stake validator and the proposal validator are only able to check for one another through the datum belonging to the ST:

    ┌─────────────────┐   ┌────────────────────┐
    │ Stake Validator ├─┐ │ Proposal Validator │
    └────────┬────────┘ │ └──────┬─────┬───────┘
             │          │        │     │
             │        ┌─┼────────┘     │
             ▼        │ │              ▼
    ┌──────────────┐  │ │ ┌─────────────────┐
    │ Stake Policy │◄─┘ └►│ Proposal Policy │
    └──────────────┘      └─────────────────┘
    

Constructors

ProposalLock

Fields

vote :: ResultTag

What was voted on. This allows retracting votes to undo their vote.

proposalId :: ProposalId

Identifies the proposal. See ProposalId for further comments on its significance.

Plutarch-land

newtype PStakeDatum (s :: S)

Plutarch-level datum for Stake scripts.

Constructors

PStakeDatum

Fields

getStakeDatum :: Term s (PDataRecord '["stakedAmount" := PDiscrete GTTag, "owner" := PPubKeyHash, "lockedBy" := PBuiltinList (PAsData PProposalLock)])

data PStakeRedeemer (s :: S)

Plutarch-level redeemer for Stake scripts.

Constructors

PDepositWithdraw (Term s (PDataRecord '["delta" := PDiscrete GTTag]))
Deposit or withdraw a discrete amount of the staked governance token.

PDestroy (Term s (PDataRecord '[]))
Destroy a stake, retrieving its LQ, the minimum ADA and any other assets.

PPermitVote (Term s (PDataRecord '["lock" := PProposalLock]))

PRetractVotes (Term s (PDataRecord '["locks" := PBuiltinList (PAsData PProposalLock)]))

PWitnessStake (Term s (PDataRecord '[]))

newtype PProposalLock (s :: S)

Plutarch-level version of ProposalLock.

Constructors

PProposalLock

Fields

getProposalLock :: Term s (PDataRecord '["vote" := PResultTag, "proposalTag" := PProposalId])

Utility functions

stakeLocked :: forall (s :: S). Term s (PStakeDatum :--> PBool)

Check whether a Stake is locked. If it is locked, various actions are unavailable.

findStakeOwnedBy :: Term s (PAssetClass :--> (PPubKeyHash :--> (PBuiltinList (PAsData (PTuple PDatumHash PDatum)) :--> (PBuiltinList (PAsData PTxInInfo) :--> PMaybe (PAsData PStakeDatum)))))

Find a stake owned by a particular PK.