Agora.Treasury

Description

Contains the datum, redeemer and validator for a template DAO treasury.

Documentation

newtype PTreasuryRedeemer (s :: S)

Plutarch level type representing valid redeemers of the treasury.

Constructors

PAlterTreasuryParams (Term s (PDataRecord '[]))
Alters treasury parameters, subject to the burning of a governance authority token.

newtype PTreasuryDatum (s :: S)

Plutarch level type representing datum of the treasury. Contains:

stateThread representing the asset class of the treasury's state thread token.

Constructors

PTreasuryDatum (Term s (PDataRecord '["stateThread" := PCurrencySymbol]))

treasuryValidator :: CurrencySymbol -> ClosedTerm PValidator

Validator ensuring that transactions consuming the treasury do so in a valid manner.