Agora.Utils

Description

Plutarch utility functions that should be upstreamed or don't belong anywhere else.

Validator-level utility functions

passert :: Term s PString -> Term s PBool -> Term s k -> Term s k

Assert a particular PBool, trace if false. Use in monadic context.

pfind' :: PIsListLike list a => (Term s a -> Term s PBool) -> Term s (list a :--> PMaybe a)

Get the first element that matches a predicate or return Nothing.

pfindDatum :: Term s (PDatumHash :--> (PBuiltinList (PAsData (PTuple PDatumHash PDatum)) :--> PMaybe PDatum))

Find a datum with the given hash.

ptryFindDatum :: forall (a :: PType) (s :: S). PTryFrom PData a => Term s (PDatumHash :--> (PBuiltinList (PAsData (PTuple PDatumHash PDatum)) :--> PMaybe a))

Find a datum with the given hash, and ptryFrom it.

pvalueSpent :: Term s (PBuiltinList (PAsData PTxInInfo) :--> PValue)

Sum of all value at input.

ptxSignedBy :: Term s (PBuiltinList (PAsData PPubKeyHash) :--> (PAsData PPubKeyHash :--> PBool))

Check if a PubKeyHash signs this transaction.

paddValue :: forall s. Term s (PValue :--> (PValue :--> PValue))

Add two PValues together.

plookup :: (PEq a, PIsListLike list (PBuiltinPair a b)) => Term s (a :--> (list (PBuiltinPair a b) :--> PMaybe b))

Find the value for a given key in an associative list.

pfromMaybe :: forall a s. Term s (a :--> (PMaybe a :--> a))

Extract a Maybe by providing a default value in case of Just.

psymbolValueOf :: Term s (PCurrencySymbol :--> (PValue :--> PInteger))

Get the sum of all values belonging to a particular CurrencySymbol.

pgeqByClass :: Term s (PCurrencySymbol :--> (PTokenName :--> (PValue :--> (PValue :--> PBool))))

Return >= on two values comparing by only a particular AssetClass.

pgeqBySymbol :: Term s (PCurrencySymbol :--> (PValue :--> (PValue :--> PBool)))

Return >= on two values comparing by only a particular CurrencySymbol.

pgeqByClass' :: AssetClass -> Term s (PValue :--> (PValue :--> PBool))

Return >= on two values comparing by only a particular Haskell-level AssetClass.

pfindTxInByTxOutRef :: Term s (PTxOutRef :--> (PBuiltinList (PAsData PTxInInfo) :--> PMaybe PTxInInfo))

Find the TxInInfo by a TxOutRef.

psingletonValue :: forall s. Term s (PCurrencySymbol :--> (PTokenName :--> (PInteger :--> PValue)))

Create a value with a single asset class.

pfindMap :: PIsListLike list a => Term s ((a :--> PMaybe b) :--> (list a :--> PMaybe b))

Get the first element that maps to a PJust in a list.

pnotNull :: forall list a. PIsListLike list a => Term _ (list a :--> PBool)

True if a list is not empty.

pisJust :: forall a s. Term s (PMaybe a :--> PBool)

Yield True if a given PMaybe is of form PJust _.

ptokenSpent :: forall {s :: S}. Term s (PAssetClass :--> (PBuiltinList (PAsData PTxInInfo) :--> PBool))

Check if a particular asset class has been spent in the input list.

When using this as an authority check, you MUST ensure the authority knows how to ensure its end of the contract.

pkeysEqual :: forall (s :: S) k a b. Term s (PMap k a :--> (PMap k b :--> PBool))

True if both maps have exactly the same keys. Using #== is not sufficient, because keys returned are not ordered.

pnub :: forall list a (s :: S). (PEq a, PIsListLike list a) => Term s (list a :--> list a)

O(n^2) . Clear out duplicates in a list. The order is not preserved.

pisUniq :: forall list a (s :: S). (PEq a, PIsListLike list a) => Term s (list a :--> PBool)

O(n^2) . Check if a list contains no duplicates.

Functions which should (probably) not be upstreamed

anyOutput :: forall (datum :: PType) s. (PIsData datum, PTryFrom PData (PAsData datum)) => Term s (PTxInfo :--> ((PValue :--> (PAddress :--> (datum :--> PBool))) :--> PBool))

Check if any output matches the predicate.

allOutputs :: forall (datum :: PType) s. (PIsData datum, PTryFrom PData (PAsData datum)) => Term s (PTxInfo :--> ((PTxOut :--> (PValue :--> (PAddress :--> (datum :--> PBool)))) :--> PBool))

Check if all outputs match the predicate.

anyInput :: forall (datum :: PType) s. (PIsData datum, PTryFrom PData (PAsData datum)) => Term s (PTxInfo :--> ((PValue :--> (PAddress :--> (datum :--> PBool))) :--> PBool))

Check if any (resolved) input matches the predicate.

findTxOutByTxOutRef :: Term s (PTxOutRef :--> (PBuiltinList (PAsData PTxInInfo) :--> PMaybe PTxOut))

Finds the TxOut of an effect from TxInfo and TxOutRef

scriptHashFromAddress :: Term s (PAddress :--> PMaybe PValidatorHash)

Get script hash from an Address.

findOutputsToAddress :: Term s (PBuiltinList (PAsData PTxOut) :--> (PAddress :--> PBuiltinList (PAsData PTxOut)))

Find all TxOuts sent to an Address

findTxOutDatum :: Term s (PBuiltinList (PAsData (PTuple PDatumHash PDatum)) :--> (PTxOut :--> PMaybe PDatum))

Find the data corresponding to a TxOut, if there is one

validatorHashToTokenName :: ValidatorHash -> TokenName

Safely convert a PValidatorHash into a PTokenName. This can be useful for tagging tokens for extra safety.

pvalidatorHashToTokenName :: forall (s :: S). Term s PValidatorHash -> Term s PTokenName

Plutarch level validatorHashToTokenName.

getMintingPolicySymbol :: ClosedTerm PMintingPolicy -> CurrencySymbol

Get the CurrencySymbol of a PMintingPolicy.