Agora.Utils.Value

Documentation

pgeq :: Term s (PValue :--> (PValue :--> PBool))

Establishes if a value is greater than or equal to another.

pleq :: Term s (PValue :--> (PValue :--> PBool))

Establishes if a value is less than or equal to another.

pgt :: Term s (PValue :--> (PValue :--> PBool))

Establishes if a value is strictly greater than another.

plt :: Term s (PValue :--> (PValue :--> PBool))

Establishes if a value is strictly less than another.