Agora.AuthorityToken

Description

Tokens acting as redeemable proofs of DAO authority.

Documentation

authorityTokenPolicy :: AuthorityToken -> Term s (PData :--> (PScriptContext :--> PUnit))

Policy given AuthorityToken params.

authorityTokensValidIn :: Term s (PCurrencySymbol :--> (PTxOut :--> PBool))

Check that all GATs are valid in a particular TxOut. How this is checked: an AuthorityToken should never leave the Effect it was initially sent to, so we simply check that the script address the token resides in matches the TokenName. Since the TokenName was tagged upon mint with the Effect script it was sent to, this is enough to prove validity.

In other words, check that all assets of a particular currency symbol are tagged with a TokenName that matches where they live.

singleAuthorityTokenBurned :: forall (s :: S). Term s PCurrencySymbol -> Term s (PAsData PTxInfo) -> Term s PValue -> Term s PBool

Assert that a single authority token has been burned.

newtype AuthorityToken

An AuthorityToken represents a proof that a particular token moved while this token was minted. In effect, this means that the validator that locked such a token must have approved said transaction. Said validator should be made aware of *this* token's existence in order to prevent incorrect minting.

Constructors

AuthorityToken

Fields

authority :: AssetClass

Token that must move in order for minting this to be valid.