Agora.Governor

Description

Governor entity scripts acting as authority of entire system.

Haskell-land

data GovernorDatum

Datum for the Governor script.

Constructors

GovernorDatum

Fields

proposalThresholds :: ProposalThresholds

Gets copied over upon creation of a ProposalDatum.

nextProposalId :: ProposalId

What tag the next proposal will get upon creating.

data GovernorRedeemer

Redeemer for Governor script. The governor has two primary responsibilities:

The gating of Proposal creation.

The gating of minting authority tokens.

Constructors

CreateProposal
Checks that a proposal was created lawfully, and allows it.

MintGATs
Checks that a SINGLE proposal finished correctly, and allows minting GATs for each effect script.

data Governor

Parameters for creating Governor scripts.

Constructors

Governor

Plutarch-land

Scripts

governorPolicy :: Governor -> ClosedTerm PMintingPolicy

Policy for Governors.

governorValidator :: Governor -> ClosedTerm PValidator

Validator for Governors.