Agora.Governor

Description

Governor entity scripts acting as authority of entire system.

Haskell-land

data GovernorDatum

Datum for the Governor script.

Constructors

GovernorDatum

Fields

proposalThresholds :: ProposalThresholds

Gets copied over upon creation of a ProposalDatum.

nextProposalId :: ProposalId

What tag the next proposal will get upon creating.

proposalTimings :: ProposalTimingConfig

The timing configuration for proposals. Will get copied over upon the creation of proposals.

createProposalTimeRangeMaxWidth :: MaxTimeRangeWidth

The maximum valid duration of a transaction that creats a proposal.

data GovernorRedeemer

Redeemer for Governor script. The governor has two primary responsibilities:

The gating of Proposal creation.

The gating of minting authority tokens.

Parameters of the governor can also be mutated by an effect.

Constructors

CreateProposal
Checks that a proposal was created lawfully, and allows it.

MintGATs
Checks that a SINGLE proposal finished correctly, and allows minting GATs for each effect script.

MutateGovernor
Allows effects to mutate the parameters.

data Governor

Parameters for creating Governor scripts.

Constructors

Governor

Fields

gstOutRef :: TxOutRef

Referenced utxo will be spent to mint the GST.

gtClassRef :: Tagged GTTag AssetClass

Governance token of the system.

maximumCosigners :: Integer

Arbitrary limit for maximum amount of cosigners on a proposal. See proposalDatumValid.

Plutarch-land

newtype PGovernorDatum (s :: S)

Plutarch-level datum for the Governor script.

Constructors

PGovernorDatum

Fields

getGovernorDatum :: Term s (PDataRecord '["proposalThresholds" := PProposalThresholds, "nextProposalId" := PProposalId, "proposalTimings" := PProposalTimingConfig, "createProposalTimeRangeMaxWidth" := PMaxTimeRangeWidth])

data PGovernorRedeemer (s :: S)

Plutarch-level version of GovernorRedeemer.

Constructors

PCreateProposal (Term s (PDataRecord '[]))

PMintGATs (Term s (PDataRecord '[]))

PMutateGovernor (Term s (PDataRecord '[]))

Utilities

pgetNextProposalId :: Term s (PProposalId :--> PProposalId)

Plutrach version of getNextProposalId.

getNextProposalId :: ProposalId -> ProposalId

Get next proposal id.

governorDatumValid :: Term s (PGovernorDatum :--> PBool)

Check whether a particular PGovernorDatum is well-formed.