Agora.Proposal

Description

Proposal scripts encoding effects that operate on the system.

Haskell-land

data Proposal

Parameters that identify the Proposal validator script.

Constructors

Proposal

Fields

governorSTAssetClass :: AssetClass

stakeSTAssetClass :: AssetClass

maximumCosigners :: Integer

Arbitrary limit for maximum amount of cosigners on a proposal.

data ProposalDatum

Haskell-level datum for Proposal scripts.

Constructors

ProposalDatum

Fields

proposalId :: ProposalId

Identification of the proposal. TODO: could we encode this more efficiently? This is shaped this way for future proofing. See https://github.com/Liqwid-Labs/agora/issues/39

effects :: Map ResultTag (Map ValidatorHash DatumHash)

Effect lookup table. First by result, then by effect hash.

status :: ProposalStatus

The status the proposal is in.

cosigners :: [PubKeyHash]

Who created the proposal initially, and who cosigned it later.

thresholds :: ProposalThresholds

Thresholds copied over on initialization.

votes :: ProposalVotes

Vote tally on the proposal

timingConfig :: ProposalTimingConfig

Timing configuration copied over on initialization.

startingTime :: ProposalStartingTime

The time upon the creation of the proposal.

data ProposalRedeemer

Haskell-level redeemer for Proposal scripts.

Constructors

Vote ResultTag
Cast one or more votes towards a particular ResultTag.

Cosign [PubKeyHash]
Add one or more public keys to the cosignature list. Must be signed by those cosigning.

This is particularly used in the Draft ProposalStatus, where matching Stakes can be called to advance the proposal, provided enough GT is shared among them.

Unlock ResultTag
Allow unlocking one or more stakes with votes towards particular ResultTag.

AdvanceProposal
Advance the proposal, performing the required checks for whether that is legal.

These are roughly the checks for each possible transition:

Draft -> VotingReady:

The sum of all of the cosigner's GT is larger than the startVoting field of ProposalThresholds.

The proposal's current time ensures isDraftPeriod.

VotingReady -> Locked:

The sum of all votes is larger than countVoting.

The winning ResultTag has more votes than all other ResultTags.

The proposal's current time ensures isVotingPeriod.

Locked -> Finished:

The proposal's current time ensures isExecutionPeriod.

The transaction mints the GATs to the receiving effects.

* -> Finished:

If the proposal has run out of time for the current ProposalStatus, it will always be possible to transition into Finished status, because it has expired (and failed).

data ProposalStatus

The "status" of the proposal. This is only useful for state transitions that need to happen as a result of a transaction as opposed to time-based "periods".

See the note on wording & the state machine in the tech-design.

If the proposal is VotingReady, for instance, that doesn't necessarily mean that voting is possible, as this also requires the timing to be right.

Constructors

Draft
A draft proposal represents a proposal that has yet to be realized.

In effect, this means one which didn't have enough LQ to be a full proposal, and needs cosigners to enable that to happen. This is similar to a "temperature check", but only useful if multiple people want to pool governance tokens together. If the proposal doesn't get to VotingReady on time, the proposal will never be able to get voted on.

VotingReady
The proposal has/had enough GT cosigned in order to be a fully fledged proposal.

This means that once the timing requirements align, proposal will be able to be voted on.

Locked
The proposal has been voted on, and the votes have been locked permanently. The proposal now goes into a locking time after the normal voting time. After this, it's possible to execute the proposal.

Finished
The proposal has finished.

This can mean it's been voted on and completed, but it can also mean the proposal failed due to time constraints or didn't get to VotingReady first.

At this stage, the votes field of ProposalDatum is frozen.

See AdvanceProposal for documentation on state transitions.

TODO: The owner of the proposal may choose to reclaim their proposal.

data ProposalThresholds

The threshold values for various state transitions to happen. This data is stored centrally (in the Governor) and copied over to Proposals when they are created.

Constructors

ProposalThresholds

Fields

execute :: Tagged GTTag Integer

How much GT minimum must a particular ResultTag accumulate for it to pass.

create :: Tagged GTTag Integer

How much GT required to "create" a proposal.

It is recommended this be a high enough amount, in order to prevent DOS from bad actors.

vote :: Tagged GTTag Integer

How much GT required to allow voting to happen. (i.e. to move into VotingReady)

newtype ProposalVotes

Map which encodes the total tally for each result. It's important that the "shape" is consistent with the shape of effects.

e.g. if the effects field looks like the following:

[(ResultTag 0, []), (ResultTag 1, [(vh, dh)])]

Then ProposalVotes needs be of the shape:

[(ResultTag 0, n), (ResultTag 1, m)]

Constructors

ProposalVotes

Fields

getProposalVotes :: Map ResultTag Integer

newtype ProposalId

Identifies a Proposal, issued upon creation of a proposal. In practice, this number starts at zero, and increments by one for each proposal. The 100th proposal will be ProposalId 99. This counter lives in the Governor. See nextProposalId, and pgetNextProposalId.

Constructors

ProposalId

Fields

proposalTag :: Integer

newtype ResultTag

Encodes a result. Typically, for a Yes/No proposal, we encode it like this:

No  ~ ResultTag 0
Yes ~ ResultTag 1

Constructors

ResultTag

Fields

getResultTag :: Integer

emptyVotesFor :: forall a. Map ResultTag a -> ProposalVotes

Create a ProposalVotes that has the same shape as the effects field.

Plutarch-land

newtype PProposalDatum (s :: S)

Plutarch-level version of ProposalDatum.

Constructors

PProposalDatum

Fields

getProposalDatum :: Term s (PDataRecord '["proposalId" := PProposalId, "effects" := PMap 'Unsorted PResultTag (PMap 'Unsorted PValidatorHash PDatumHash), "status" := PProposalStatus, "cosigners" := PBuiltinList (PAsData PPubKeyHash), "thresholds" := PProposalThresholds, "votes" := PProposalVotes, "timingConfig" := PProposalTimingConfig, "startingTime" := PProposalStartingTime])

data PProposalRedeemer (s :: S)

Plutarch-level version of ProposalRedeemer.

Constructors

PVote (Term s (PDataRecord '["resultTag" := PResultTag]))

PCosign (Term s (PDataRecord '["newCosigners" := PBuiltinList (PAsData PPubKeyHash)]))

PUnlock (Term s (PDataRecord '["resultTag" := PResultTag]))

PAdvanceProposal (Term s (PDataRecord '[]))

data PProposalStatus (s :: S)

Plutarch-level version of ProposalStatus.

Constructors

PDraft (Term s (PDataRecord '[]))

PVotingReady (Term s (PDataRecord '[]))

PLocked (Term s (PDataRecord '[]))

PFinished (Term s (PDataRecord '[]))

newtype PProposalThresholds (s :: S)

Plutarch-level version of ProposalThresholds.

Constructors

PProposalThresholds

Fields

getProposalThresholds :: Term s (PDataRecord '["execute" := PDiscrete GTTag, "create" := PDiscrete GTTag, "vote" := PDiscrete GTTag])

newtype PProposalVotes (s :: S)

Plutarch-level version of ProposalVotes.

Constructors

PProposalVotes (Term s (PMap 'Unsorted PResultTag PInteger))

newtype PProposalId (s :: S)

Plutarch-level version of PProposalId.

Constructors

PProposalId (Term s PInteger)

newtype PResultTag (s :: S)

Plutarch-level version of ResultTag.

Constructors

PResultTag (Term s PInteger)

Plutarch helpers

proposalDatumValid :: Proposal -> Term s (PProposalDatum :--> PBool)

Check for various invariants a proposal must uphold. This can be used to check both upon creation and upon any following state transitions in the proposal.

pemptyVotesFor :: forall s a. PIsData a => Term s (PMap 'Unsorted PResultTag a :--> PProposalVotes)

Plutarch-level version of emptyVotesFor.

pwinner :: Term s (PProposalVotes :--> (PInteger :--> (PResultTag :--> PResultTag)))

Find the winner result tag, given the votes, the quorum the "neutral" result tag.

The winner should be unambiguous, meaning that if two options have the same highest votes, the "neutral" option will be the winner.

pneutralOption :: Term s (PMap 'Unsorted PResultTag (PMap 'Unsorted PValidatorHash PDatumHash) :--> PResultTag)

Find the "neutral" option (a dummy outcome with no effect) given the effects.

pretractVotes :: Term s (PResultTag :--> (PInteger :--> (PProposalVotes :--> PProposalVotes)))

Retract votes given the option and the amount of votes.