Agora.Treasury

Description

Contains the datum, redeemer and validator for a template DAO treasury.

Documentation

data TreasuryRedeemer

Redeemer for Treasury actions.

Constructors

SpendTreasuryGAT
Allow transaction to pass by delegating to GAT burn.

newtype PTreasuryRedeemer (s :: S)

Plutarch level type representing valid redeemers of the treasury.

Constructors

PSpendTreasuryGAT (Term s (PDataRecord '[]))
Alters treasury parameters, subject to the burning of a governance authority token.

treasuryValidator

Arguments

:: CurrencySymbol
Governance Authority Token that can unlock this validator.

-> ClosedTerm PValidator

Validator ensuring that transactions consuming the treasury do so in a valid manner.