Agora.Utils

Description

Plutarch utility functions that should be upstreamed or don't belong anywhere else.

Documentation

findTxOutByTxOutRef :: Term s (PTxOutRef :--> (PBuiltinList (PAsData PTxInInfo) :--> PMaybe PTxOut))

Finds the TxOut of an effect from TxInfo and TxOutRef

scriptHashFromAddress :: Term s (PAddress :--> PMaybe PValidatorHash)

Get script hash from an Address.

findOutputsToAddress :: Term s (PBuiltinList (PAsData PTxOut) :--> (PAddress :--> PBuiltinList (PAsData PTxOut)))

Find all TxOuts sent to an Address

findTxOutDatum :: Term s (PBuiltinList (PAsData (PTuple PDatumHash PDatum)) :--> (PTxOut :--> PMaybe PDatum))

Find the data corresponding to a TxOut, if there is one

validatorHashToTokenName :: ValidatorHash -> TokenName

Safely convert a PValidatorHash into a PTokenName. This can be useful for tagging tokens for extra safety.

pvalidatorHashToTokenName :: forall (s :: S). Term s PValidatorHash -> Term s PTokenName

Plutarch level validatorHashToTokenName.

getMintingPolicySymbol :: ClosedTerm PMintingPolicy -> CurrencySymbol

Get the CurrencySymbol of a PMintingPolicy.

hasOnlyOneTokenOfCurrencySymbol :: forall (keys :: KeyGuarantees) (amounts :: AmountGuarantees) (s :: S). Term s (PCurrencySymbol :--> (PValue keys amounts :--> PBool))

The entire value only contains one token of the given currency symbol.

mustFindDatum' :: forall (datum :: PType). (PIsData datum, PTryFrom PData (PAsData datum)) => forall s. Term s (PMaybeData PDatumHash :--> (PBuiltinList (PAsData (PTuple PDatumHash PDatum)) :--> datum))

Find datum given a maybe datum hash

mustBePJust :: forall a s. Term s (PString :--> (PMaybe a :--> a))

Extract the value stored in a PMaybe container. If there's no value, throw an error with the given message.

mustBePDJust :: forall a s. PIsData a => Term s (PString :--> (PMaybeData a :--> a))

Extract the value stored in a PMaybeData container. If there's no value, throw an error with the given message.

validatorHashToAddress :: ValidatorHash -> Address

Create an Address from a given ValidatorHash with no StakingCredential.

isScriptAddress :: Term s (PAddress :--> PBool)

Return true if the given address is a script address.

isPubKey :: Term s (PCredential :--> PBool)

Return true if the given credential is a pub-key-hash.

psingletonValue :: forall (keys :: KeyGuarantees) (amounts :: AmountGuarantees) (s :: S). Term s (PCurrencySymbol :--> (PTokenName :--> (PInteger :--> PValue keys amounts)))

Create a value with a single asset class.